modify map and cmd

+++ src/navi_manage.cpp
@@ -670,7 +670,7 @@ int HeadingAnalysis(int Heading,int Bearing)
 	 DEBUG(LOG_DEBUG,"targetPosition: %d ,meters:%d \n",targetPosition,meters);
   do{
   	 
-	 done = dwa_loop(meters);
+	 done = dwa_loop((float)meters);
 
      usleep(10000);
      }
